Refactor kernel to support x86_64 architecture and implement multiheap memory management
- Updated process loading to use 64-bit instruction pointers (RIP) instead of 32-bit (IP).
- Converted assembly code to 64-bit mode, adjusting register usage and stack operations.
- Changed task structure to utilize 64-bit registers for task state management.
- Modified memory management functions to implement a multiheap system, allowing for dynamic memory allocation across multiple heaps.
- Introduced new multiheap API for creating, adding, and managing multiple heaps.
- Updated build configuration to target x86_64 architecture, including toolchain and assembly format changes.
- Added documentation for EDK2/OVMF integration and usage instructions for UEFI booting.

+++ include/status.h
@@ -12,7 +12,7 @@
  * - <0 : error values encoded as -errno
  */
 
-typedef int32_t status_t;
+typedef int64_t status_t;
 
 /* =========================
  * Standard error codes
