+++ include/stdint.h
@@ -13,8 +13,8 @@ typedef unsigned int uint32_t;
 typedef signed long long int64_t;
 typedef unsigned long long uint64_t;
 
-typedef unsigned int size_t;
-typedef signed int intptr_t;
-typedef unsigned int uintptr_t;
+typedef unsigned long long size_t;
+typedef signed long long intptr_t;
+typedef unsigned long long uintptr_t;
 
 #endif /* STDINT_H */
